[app.tsx edit]

+++ src/components/AllBook/AllBook.tsx
@@ -58,10 +58,16 @@ const AllBookComponent: React.FC = () => {
   };
 
   const deleteBookHandler = (id: string) => {
-    deleteBook(id);
-    toast.success("Book deleted", {
-      id: "delete-book",
-    });
+    const shouldDelete = window.confirm(
+      "Are you sure you want to delete this book?"
+    );
+
+    if (shouldDelete) {
+      deleteBook(id);
+      toast.success("Book deleted", {
+        id: "delete-book",
+      });
+    }
   };
 
   let content;
